Doc/Clean: Added comments in transform.processed_results.main.get_processed_results and removed useless astype call.

+++ src/transform/processed_results/main.py
@@ -54,7 +54,9 @@ def get_processed_results(brand:str) -> DF:
     logger.info(f"{'Processing ' + brand + ' results.':=^{50}}")
     results =  (
         GET_RESULTS_FUNCS[brand]()
-        .assign(soh=lambda df: df["soh"].replace([np.inf, -np.inf], np.nan).astype("float64"))
+        # Some raw estimations may have inf values, this will make mask_out_outliers_by_interquartile_range and force_monotonic_decrease fail
+        # So we replace them by NaNs.
+        .assign(soh=lambda df: df["soh"].replace([np.inf, -np.inf], np.nan))
         .sort_values(["vin", "date"])
         .pipe(agg_results_by_update_frequency)
         .pipe(make_charge_levels_presentable)
